work: fixing x32 test_workpool_basic

+++ solid/utility/test/test_workpool_basic.cpp
@@ -19,13 +19,13 @@ int test_workpool_basic(int /*argc*/, char* /*argv*/ [])
     using WorkPoolT  = WorkPool<size_t>;
     using AtomicPWPT = std::atomic<WorkPoolT*>;
 
-    const int           wait_seconds = 500;
-    const int           loop_cnt = 10;
-    const size_t        cnt{5000000};
-    std::atomic<size_t> val{0};
-    promise<void>       prom;
-    AtomicPWPT          pwp{nullptr};
-    thread              wait_thread(
+    const int             wait_seconds = 500;
+    const int             loop_cnt     = 10;
+    const uint64_t        cnt{5000000};
+    std::atomic<uint64_t> val{0};
+    promise<void>         prom;
+    AtomicPWPT            pwp{nullptr};
+    thread                wait_thread(
         [](promise<void>& _rprom, AtomicPWPT& _rpwp, const int _wait_time_seconds) {
             if (_rprom.get_future().wait_for(chrono::seconds(_wait_time_seconds)) != future_status::ready) {
                 if (_rpwp != nullptr) {
@@ -36,7 +36,7 @@ int test_workpool_basic(int /*argc*/, char* /*argv*/ [])
         },
         std::ref(prom), std::ref(pwp), wait_seconds);
 
-    for(int i = 0; i < loop_cnt; ++i){
+    for (int i = 0; i < loop_cnt; ++i) {
         WorkPoolT wp{
             2,
             WorkPoolConfiguration(),
@@ -56,7 +56,7 @@ int test_workpool_basic(int /*argc*/, char* /*argv*/ [])
     wait_thread.join();
     solid_log(logger, Verbose, "after wait thread join: val = " << val);
 
-    const size_t v = (loop_cnt * ((cnt - 1) * cnt)) / 2;
+    const uint64_t v = (loop_cnt * ((cnt - 1) * cnt)) / 2;
 
     solid_check(v == val);
     return 0;
